+++ app/controllers/domains_controller.rb
@@ -101,6 +101,8 @@ class DomainsController < ApplicationController
       # Always update user_id to correctly track sheet transactions
       params[:domain][:user_id] = current_user.id # unless @domain
 
+      params[:domain] = Domain.clean_option_tokens(params[:domain])
+
       params.require(:domain).permit(
         :name, :display_name, :description, :user_id, { :option_tokens => [ :name, :value, :description, :missing_code, :option_index ] }
       )