+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   - Mass importing sheets now notifies the user doing the import instead of the user who created the design
   - Existing variables can now be added to a design without needing to select the type of question first
   - Public surveys have a setting to allow users to select a site while filling out a survey
+- **Domain Changes**
+  - Option values no longer need to be entered and will fall back to the default placeholder numbers if left blank
 - **Gem Changes**
   - Updated to rails 4.2.0.beta1
   - Updated to contour 2.6.0.beta6